Updated dependency management and removed unnecessary files
This is start of a rather radical change set; teanity will get modularized so everybody can pick only pieces of this library in order to build what they need without picking up unnecessary libraries along the way.
Also app module will be moved to another repository as it doesn't really makes sense to keep it here anymore. It will be easier to copy and start from this way.