Added CompoundUseCase
It's meant as a new successor of regular UseCase and LiveUseCase.

The intention is to bring together the best of both worlds, you'll encounter countless cases where you only have LiveUseCase bound to Room (or other observable source) which is not reliable source of immediate truth.
CompoundUseCase allows LiveUseCase to join UseCase in that matter. The usecase performs regular operations in suspense and allows its observe method to be overridden and replaced by observable source instead. This way you can fetch data "synchronously" and observe them in the same place (although from different source methods).