Added ActivityContractHelper
TL;DR: you can now request permissions and await activity results in viewModels.

Added new androidx activity dependency that allows using contracts as part of the activity itself. The ViewModel required another helper to wrap the contracts and ship them to the views, so ActivityContractHelper was added to achieve that.
Permissions extensions have been marked as deprecated and will be removed on 1.2.